# Use AVX2 to speedup matmulQ40