@b4rtaz 2024-05-15
Confirmed the speed up.

Setup: GitHub codepaces, 4-core AMD EPYC 7763 64-Core Processor, 16GB RAM

0.5.0

```
@b4rtaz ➜ /workspaces/distributed-llama (main) $ ./main inference --model ./dllama_meta-llama-3-8b_q40.bin --tokenizer dllama_meta-llama3-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4
...
⏩ Loaded 6175568 kB
🔶 G  755 ms I  755 ms T    0 ms S      0 kB R      0 kB Hello
🔶 G  730 ms I  730 ms T    0 ms S      0 kB R      0 kB  world
🔶 G  759 ms I  758 ms T    0 ms S      0 kB R      0 kB !
🔶 G  819 ms I  811 ms T    7 ms S      0 kB R      0 kB  <
🔶 G  717 ms I  715 ms T    1 ms S      0 kB R      0 kB br
🔶 G  874 ms I  862 ms T   11 ms S      0 kB R      0 kB >

🔶 G  710 ms I  708 ms T    0 ms S      0 kB R      0 kB  <
🔶 G  833 ms I  827 ms T    5 ms S      0 kB R      0 kB h
🔶 G  764 ms I  762 ms T    1 ms S      0 kB R      0 kB 1
🔶 G  726 ms I  725 ms T    0 ms S      0 kB R      0 kB  align
🔶 G  864 ms I  857 ms T    6 ms S      0 kB R      0 kB ="
🔶 G  813 ms I  808 ms T    4 ms S      0 kB R      0 kB center
🔶 G  698 ms I  697 ms T    0 ms S      0 kB R      0 kB ">
🔶 G  746 ms I  739 ms T    6 ms S      0 kB R      0 kB Hi
🔶 G  710 ms I  709 ms T    0 ms S      0 kB R      0 kB  �
🔶 G  717 ms I  714 ms T    2 ms S      0 kB R      0 kB 
Generated tokens:    16
Avg tokens / second: 1.31
Avg generation time: 764.69 ms
Avg inference time:  761.06 ms
Avg transfer time:   2.69 ms
```

Your PR:

```
@b4rtaz ➜ /workspaces/distributed-llama (main) $ ./main inference --model ./dllama_meta-llama-3-8b_q40.bin --tokenizer dllama_meta-llama3-tokenizer.t --weights-float-type q40 --buffer-float-type q80 --prompt "Hello world" --steps 16 --nthreads 4
...
⏩ Loaded 6175568 kB
🔶 G  568 ms I  567 ms T    1 ms S      0 kB R      0 kB Hello
🔶 G  642 ms I  642 ms T    0 ms S      0 kB R      0 kB  world
🔶 G  579 ms I  578 ms T    0 ms S      0 kB R      0 kB !
🔶 G  566 ms I  565 ms T    0 ms S      0 kB R      0 kB  


🔶 G  646 ms I  643 ms T    1 ms S      0 kB R      0 kB I
🔶 G  563 ms I  562 ms T    0 ms S      0 kB R      0 kB  am
🔶 G  818 ms I  785 ms T   32 ms S      0 kB R      0 kB  a
🔶 G  593 ms I  585 ms T    7 ms S      0 kB R      0 kB  computer
🔶 G  761 ms I  737 ms T   23 ms S      0 kB R      0 kB  science
🔶 G  579 ms I  579 ms T    0 ms S      0 kB R      0 kB  student
🔶 G  566 ms I  564 ms T    1 ms S      0 kB R      0 kB  in
🔶 G  625 ms I  623 ms T    0 ms S      0 kB R      0 kB  China
🔶 G  618 ms I  616 ms T    1 ms S      0 kB R      0 kB .
🔶 G  573 ms I  573 ms T    0 ms S      0 kB R      0 kB  I
🔶 G  690 ms I  657 ms T   32 ms S      0 kB R      0 kB  have
🔶 G  668 ms I  646 ms T   22 ms S      0 kB R      0 kB  been
Generated tokens:    16
Avg tokens / second: 1.59
Avg generation time: 628.44 ms
Avg inference time:  620.12 ms
Avg transfer time:   7.50 ms
```